requirements.txt
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
matplotlib==3.7.1